Moved MapboxExtendableButton to libnavui-base module. Added libnavui-resources as implementation depedency to libnavui-base. Added added androidx AppCompat and ConstraintLayout dependencies to libnavui-base. Moved MapboxCameraModeButton to livnavui-maps Updated Metalava API files for libnavui-base, libnavui-dropin and libnavui-maps modules.
