+++ libnavui-base/build.gradle
@@ -20,6 +20,10 @@ android {
         targetCompatibility = JavaVersion.VERSION_1_8
     }
 
+    buildFeatures {
+        viewBinding true
+    }
+
     defaultConfig {
         minSdkVersion androidVersions.minSdkVersion
         targetSdkVersion androidVersions.targetSdkVersion
@@ -37,11 +41,14 @@ dependencies {
     api project(':libnavigation-base')
     api project(':libnavigation-core')
     api project(":libnavui-util")
+    implementation project(":libnavui-resources")
     implementation dependenciesList.kotlinStdLib
     api dependenciesList.mapboxSdkGeoJSON
     api dependenciesList.androidXCore
     api dependenciesList.androidXAnnotation
     api dependenciesList.viewBinding
+    implementation dependenciesList.androidXAppCompat
+    implementation dependenciesList.androidXConstraintLayout
 
     apply from: "${rootDir}/gradle/unit-testing-dependencies.gradle"
     testImplementation(project(':libtesting-utils'))
