+++ libnavui-base/src/main/res/values/attrs.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <declare-styleable name="MapboxExtendableButton">
+        <!-- Background to apply to the view. -->
+        <attr name="extendableButtonBackground" format="reference|color" />
+        <!-- Default icon drawable -->
+        <attr name="extendableButtonIcon" format="reference" />
+        <!-- Tint to apply to the icon drawable. -->
+        <attr name="extendableButtonIconTint" format="reference|color" />
+        <!-- Appearance of a text message. -->
+        <attr name="extendableButtonTextAppearance" format="reference" />
+    </declare-styleable>
+</resources>
