+++ libnavui-dropin/src/main/res/layout/mapbox_info_panel_header_layout.xml
@@ -50,7 +50,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.5" />
 
-        <com.mapbox.navigation.dropin.view.MapboxExtendableButton
+        <com.mapbox.navigation.ui.base.view.MapboxExtendableButton
             android:id="@+id/endNavigation"
             android:layout_width="72dp"
             android:layout_height="52dp"
@@ -75,7 +75,7 @@
             app:layout_constraintBottom_toBottomOf="@+id/routePreview"
             app:layout_constraintEnd_toStartOf="@+id/routePreview" />
 
-        <com.mapbox.navigation.dropin.view.MapboxExtendableButton
+        <com.mapbox.navigation.ui.base.view.MapboxExtendableButton
             android:id="@+id/routePreview"
             android:layout_width="72dp"
             android:layout_height="52dp"
@@ -84,7 +84,7 @@
             app:layout_constraintBottom_toBottomOf="@+id/startNavigation"
             app:layout_constraintTop_toTopOf="@+id/startNavigation" />
 
-        <com.mapbox.navigation.dropin.view.MapboxExtendableButton
+        <com.mapbox.navigation.ui.base.view.MapboxExtendableButton
             android:id="@+id/startNavigation"
             android:layout_width="72dp"
             android:layout_height="52dp"
