+++ libnavui-dropin/src/main/res/values/dimens.xml
@@ -17,13 +17,6 @@
     <dimen name="mapbox_actionList_width">76dp</dimen>
     <dimen name="mapbox_actionList_spacing">5dp</dimen>
 
-    <!-- MapboxExtendableButton -->
-    <dimen name="mapbox_extendableButton_paddingStart">10dp</dimen>
-    <dimen name="mapbox_extendableButton_minHeight">52dp</dimen>
-    <dimen name="mapbox_extendableButton_minWidth">72dp</dimen>
-    <dimen name="mapbox_extendableButton_iconSize">32dp</dimen>
-    <dimen name="mapbox_extendableButton_spacing">8dp</dimen>
-
     <!-- Maneuver view -->
     <dimen name="mapbox_maneuver_view_padding">4dp</dimen>
