+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/component/cameramode/CameraModeButtonComponentTest.kt
@@ -6,12 +6,12 @@ import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.dropin.R
 import com.mapbox.navigation.dropin.util.TestStore
-import com.mapbox.navigation.dropin.view.MapboxCameraModeButton
 import com.mapbox.navigation.testing.MainCoroutineRule
 import com.mapbox.navigation.ui.app.internal.State
 import com.mapbox.navigation.ui.app.internal.camera.TargetCameraMode
 import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.maps.camera.state.NavigationCameraState
+import com.mapbox.navigation.ui.maps.view.MapboxCameraModeButton
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.just
