+++ libnavui-maps/src/main/res/drawable/mapbox_ic_camera_follow.xml
