+++ libnavui-maps/src/main/res/drawable/mapbox_ic_camera_overview.xml
