+++ libnavui-maps/src/main/res/layout/mapbox_camera_mode_button_layout.xml
