+++ libnavui-maps/src/main/res/values/styles.xml
@@ -6,4 +6,41 @@
     <style name="MapboxStyleRouteOverview">
         <item name="overviewButtonDrawable">@drawable/mapbox_ic_route_overview</item>
     </style>
+
+    <!-- Camera mode button -->
+    <style name="MapboxCameraModeButtonTextAppearance">
+        <item name="android:textColor">@color/colorOnSurface</item>
+        <item name="android:textSize">@dimen/mapbox_button_text_regular</item>
+    </style>
+
+    <style name="MapboxStyleCameraModeButton">
+        <item name="android:elevation">0dp</item>
+        <item name="android:minHeight">@dimen/mapbox_extendableButton_minHeight</item>
+        <item name="android:paddingStart">@dimen/mapbox_extendableButton_paddingStart</item>
+        <item name="android:paddingEnd">@dimen/mapbox_extendableButton_paddingStart</item>
+        <item name="cameraModeButtonBackground">@drawable/mapbox_bg_button</item>
+        <item name="cameraModeButtonIconTint">@color/colorOnSurface</item>
+        <item name="cameraModeButtonFollowIcon">@drawable/mapbox_ic_camera_follow</item>
+        <item name="cameraModeButtonOverviewIcon">@drawable/mapbox_ic_camera_overview</item>
+        <item name="cameraModeButtonTextAppearance">
+            @style/MapboxCameraModeButtonTextAppearance
+        </item>
+        <item name="cameraModeButtonFollowText">@string/mapbox_route_follow</item>
+        <item name="cameraModeButtonOverviewText">@string/mapbox_route_overview</item>
+    </style>
+
+    <style name="MapboxStyleCameraModeButton.Oval" parent="MapboxStyleCameraModeButton">
+        <item name="cameraModeButtonBackground">@drawable/mapbox_bg_circle_outline</item>
+    </style>
+
+    <style name="MapboxStyleCameraModeButton.Square" parent="MapboxStyleCameraModeButton">
+        <item name="android:paddingStart">0dp</item>
+        <item name="android:paddingEnd">0dp</item>
+        <item name="android:minHeight">0dp</item>
+    </style>
+
+    <style name="MapboxStyleCameraModeButton.Circle" parent="MapboxStyleCameraModeButton.Square">
+        <item name="cameraModeButtonBackground">@drawable/mapbox_bg_circle_outline</item>
+    </style>
+
 </resources>
