+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/DropInButtonsActivity.kt
@@ -6,8 +6,8 @@ import android.view.View
 import androidx.lifecycle.MutableLiveData
 import com.mapbox.maps.CameraOptions
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
-import com.mapbox.navigation.dropin.view.MapboxCameraModeButton
-import com.mapbox.navigation.dropin.view.MapboxExtendableButton
+import com.mapbox.navigation.ui.maps.view.MapboxCameraModeButton
+import com.mapbox.navigation.ui.base.view.MapboxExtendableButton
 import com.mapbox.navigation.qa_test_app.R
 import com.mapbox.navigation.qa_test_app.databinding.LayoutActivityDropinButtonsBinding
 import com.mapbox.navigation.qa_test_app.view.base.BaseNavigationActivity
